Suppression de la classe Action. Classe Personnage : ajout de méthodes+exception pour chaque action possible pour un personnage pendant un tour. Classe jeu : ajout d'une méthode tour qui sert d'interface de joue pour q'un joueur face son tour, modification de la méthode start. Classe arene : ajout d'une méthode qui affiche les information/stats des personnages (nbParade, energie, pv,...)

+++ src/attaques/Attaques.java
@@ -3,21 +3,21 @@ package attaques;
 public class Attaques {
 
     private String type_attaque;
-    private int degat;
+    private double degat;
     private int portee;
     
 
-    public Attaques(String type_attaque, int degat, int portee){
+    public Attaques(String type_attaque, double degat, int portee){
         this.type_attaque = type_attaque;
         this.degat = degat;
         this.portee = portee;
     }
 
-    public int getDegat(){
+    public double getDegat(){
         return degat;
     }
 
-    public void setDegat(int supplement){
+    public void setDegat(double supplement){
         this.degat += supplement;
     }
